requirements.txt
alembic==1.9.1
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
cli-helpers==2.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configobj==5.0.8
cryptography==39.0.0
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.89.1
filelock==3.9.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
identify==2.5.15
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.24.1
orjson==3.8.5
packaging==23.0
pandas==1.5.3
pendulum==2.1.2
pgcli==3.5.0
pgspecial==2.0.1
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.1
prompt-toolkit==3.0.36
psycopg==3.1.8
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
redis==4.4.2
rfc3986==1.5.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
sqlalchemy2-stubs==0.0.2a32
sqlparse==0.4.3
starlette==0.22.0
tabulate==0.9.0
tomli==2.0.1
types-pyOpenSSL==23.0.0.2
types-redis==4.4.0.4
typing_extensions==4.4.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
virtualenv==20.17.1
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
xlwt==1.3.0
